Move `OffsetStoreRepository` tests into `providertest`.
